refactor(gateway): split publication owners
